requirements.txt
django-debug-toolbar
django-bootstrap5
django-htmx
pandas
matplotlib